Extract script creations to createScripts
